refactor: simplify user handling by consolidating user struct and removing bot/person enums
