+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
+  "editor.formatOnSave": true,
   "rust-analyzer.linkedProjects": ["./Cargo.toml"],
   "rust-analyzer.showUnlinkedFileNotification": false
 }
