+++ notionrs/examples/user_list_users.rs
@@ -7,10 +7,7 @@ async fn main() -> Result<(), notionrs::error::Error> {
     let response = request.send().await?;
 
     for user in response.results {
-        let name = match user {
-            notionrs::object::user::User::Bot(bot) => bot.name,
-            notionrs::object::user::User::Person(person) => person.name,
-        };
+        let name = user.name;
 
         match name {
             None => println!("No name found"),
